requirements.txt
fastapi
uvicorn
pymupdf
python-multipart
jinja2
sarvamai
python-dotenv
pytesseract
pillow
